+++ src/BusBundle/Controller/BusTypeController.php
@@ -4,8 +4,11 @@ namespace BusBundle\Controller;
 
 use ApiLogsBundle\Entity\BusTypeTableLog;
 use BusBundle\Entity\BusType;
+use BusBundle\Repository\BusTypeRepository;
 use Extra\ApiClient;
+use Extra\ApiProblem;
 use Extra\DateTimeProvider;
+use Extra\ResponseFactory;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Method;
@@ -18,6 +21,9 @@ class BusTypeController extends Controller
     /**
      * @Route("bus-types/{id}")
      * @Method("PUT")
+     * @param $id
+     * @param $request
+     * @return JsonResponse
      */
     public function editBusTypeAction($id, Request $request)
     {
@@ -30,9 +36,7 @@ class BusTypeController extends Controller
         if (ApiClient::hasAdminCredentialsForJSONRequest($username)) {
             $em = $this->getDoctrine()->getManager();
 
-            $busTypeEntity = $this->getDoctrine()
-                ->getRepository('BusBundle:BusType')
-                ->find($id);
+            $busTypeEntity = $this->getBusTypeRepository()->find($id);
 
             $originalBusType = $busTypeEntity->getType();
 
@@ -63,68 +67,56 @@ class BusTypeController extends Controller
         return new JsonResponse($response);
     }
 
-    /**
-     * @Route("/check/bus-type-existence")
-     * @Method("POST")
-     * 
-     * Used to check if a bus already exists based on the user input for the 
-     * busType field
-     */
-    public function checkBusTypeExistenceAction(Request $request)
-    {
-        $data = json_decode($request->getContent(), true);
-
-        $busTypes = $this->getDoctrine()
-            ->getRepository("BusBundle:BusType")
-            ->findAll();
-
-        $busTypeExists = false;
-
-        foreach ($busTypes as $busType)
-        {
-            if (strcasecmp($busType->getType(), $data['type']) === 0)
-            {
-                $busTypeExists = true;
-                break;
-            }
-        }
-
-        $response = [
-            'busTypeExists' => $busTypeExists
-        ];
-
-        return new JsonResponse($response);
-    }
 
     /**
      * @Route("/bus-types/{id}")
      * @Method("DELETE")
+     * @param $id
+     * @param $request
+     * @return JsonResponse
      */
     public function deleteBusTypeAction($id, Request $request)
     {
         $em = $this->getDoctrine()->getManager();
-        $busType = $em->getRepository('BusBundle:BusType')->find($id);
 
-        $busTypeToBeRemoved = $busType->getType();
+        if ($busType = $this->getBusTypeRepository()->find($id)) {
 
-        $busTypeTableLogEntity = new BusTypeTableLog();
-        $busTypeTableLogEntity->setAction("Removed $busTypeToBeRemoved");
-        $busTypeTableLogEntity->setCreated(DateTimeProvider::getNow());
-        $busTypeTableLogEntity->setClientIP($request->getClientIp());
-        $busTypeTableLogEntity->setUser('awesome_admin');
+            $busTypeToBeRemoved = $busType->getType();
 
-        $em->remove($busType);
-        $em->persist($busTypeTableLogEntity);
-        $em->flush();
+            $busTypeTableLogEntity = new BusTypeTableLog();
+            $busTypeTableLogEntity->setAction("Removed $busTypeToBeRemoved");
+            $busTypeTableLogEntity->setCreated(DateTimeProvider::getNow());
+            $busTypeTableLogEntity->setClientIP($request->getClientIp());
+            $busTypeTableLogEntity->setUser('awesome_admin');
+
+            $em->remove($busType);
+            $em->persist($busTypeTableLogEntity);
+            $em->flush();
+
+            $response = [
+                'message' => 'Successfully removed item.',
+            ];
+
+            return new JsonResponse($response, 200);
+        }
         
-        $response = [
-            'message' => 'Successfully removed item.',
-            'status' => 'Success'
-        ];
-        
-        return new JsonResponse($response);
+        $responseFactory = new ResponseFactory();
+        $apiProblem = new ApiProblem(404);
+        $apiProblem->set('detail', 'Resource not found');
+
+        return $responseFactory->createResponse($apiProblem);
+
     }
-    
+
+    /**
+     * @return BusTypeRepository
+     */
+    public function getBusTypeRepository()
+    {
+        return $this->getDoctrine()
+            ->getRepository("BusBundle:BusType");
+    }
+
     /**
      * @Route("/bus-types")
      * @Method("POST")
@@ -132,43 +124,75 @@ class BusTypeController extends Controller
     public function createBusTypeAction(Request $request)
     {
         $data = json_decode($request->getContent(), true);
+        $errors = [];
+        $busTypeRepository = $this->getBusTypeRepository();
 
-        $busType = $data['type'];
-        $description = $data['description'];
-        $username = $data['username'];
-        
-        if (ApiClient::hasAdminCredentialsForJSONRequest($username)) {
-            
-            $busTypeEntity = new BusType();
-
-            $busTypeEntity->setType($busType);
-            $busTypeEntity->setDescription($description);
-            $busTypeEntity->setCreated(DateTimeProvider::getNow());
-            $busTypeEntity->setUpdated(DateTimeProvider::getNow());
-
-            $busTypeTableLogEntity = new BusTypeTableLog();
-            $busTypeTableLogEntity->setClientIP($request->getClientIp());
-            $busTypeTableLogEntity->setUser($username);
-            $busTypeTableLogEntity->setCreated(DateTimeProvider::getNow());
-            $busTypeTableLogEntity->setAction("Added $busType");
-            
-            $em = $this->getDoctrine()->getManager();
-            $em->persist($busTypeEntity);
-            $em->persist($busTypeTableLogEntity);
-            $em->flush();
-
-            $response = [
-                'message' => $busType . " successfully added as a bus type.",
-                'status' => 'success'
-            ];
-        } else {
-            $response = [
-                'message' => "Insufficient permissions",
-                'status' => 'error'
-            ];
+        if (!$busType = $data['type']) {
+            $errors[] = '"Bus Type" is required';
         }
 
-        return new JsonResponse($response);
+        if (!$description = $data['description']) {
+            $errors[] = '"Description" is required';
+        }
+
+        if (!$username = $data['username']) {
+            $errors[] = '"username" is required';
+        } else {
+            if (ApiClient::hasAdminCredentialsForJSONRequest($username) && count($errors) === 0) {
+
+                if ($busType = $busTypeRepository->findRecordByBusType($busType)) {
+                    $response = [
+                        'message' => $data['type'] . " already exists as a bus type."
+                    ];
+                    $status = 400;
+
+                    return new JsonResponse($response, $status);
+                } else {
+
+                    // All is well, let's create the record
+                    $busTypeEntity = new BusType();
+
+                    $busTypeEntity->setType($data['type']);
+                    $busTypeEntity->setDescription($description);
+                    $busTypeEntity->setCreated(DateTimeProvider::getNow());
+                    $busTypeEntity->setUpdated(DateTimeProvider::getNow());
+
+                    $busTypeTableLogEntity = new BusTypeTableLog();
+                    $busTypeTableLogEntity->setClientIP($request->getClientIp());
+                    $busTypeTableLogEntity->setUser($username);
+                    $busTypeTableLogEntity->setCreated(DateTimeProvider::getNow());
+                    $busTypeTableLogEntity->setAction("Added $busType");
+
+                    $em = $this->getDoctrine()->getManager();
+                    $em->persist($busTypeEntity);
+                    $em->persist($busTypeTableLogEntity);
+                    $em->flush();
+
+                    $response = [
+                        'message' => $data['type'] . " successfully added as a bus type."
+                    ];
+
+                    $status = 200;
+                    return new JsonResponse($response, $status);
+                }
+            } else {
+                if (!ApiClient::hasAdminCredentialsForJSONRequest($username)) {
+                    $responseFactory = new ResponseFactory();
+                    $apiProblem = new ApiProblem(403);
+                    $apiProblem->set('detail', 'Access denied');
+
+                    return $responseFactory->createResponse($apiProblem);
+                }
+
+                if (count($errors) != 0) {
+                    $response = [
+                        'message' => 'There were some issues creating the Bus Type',
+                        'errors' => $errors
+                    ];
+                    return new JsonResponse($response, 400);
+                }
+            }
+        }
     }
 
     /**
@@ -177,21 +201,18 @@ class BusTypeController extends Controller
      */
     public function listBusTypeAction()
     {
-        $busTypes = $this->getDoctrine()
-            ->getRepository("BusBundle:BusType")
-            ->findAll();
-        
+        $busTypes = $this->getBusTypeRepository()->findAll();
+
         $response = [];
-        
-        foreach ($busTypes as $busType)
-        {
+
+        foreach ($busTypes as $busType) {
             $response[] = [
                 'id' => $busType->getId(),
                 'type' => $busType->getType(),
                 'description' => $busType->getDescription()
             ];
         }
-        
+
         return new JsonResponse($response);
     }
 }
